{"schema":"decodex/commit/1","summary":"Modularize recovery evidence and run capsules","authority":"manual"}
